Home page product list added
