+++ src/main/java/com/controller/Cart/PlusCartItemQuantity.java
@@ -1,4 +1,4 @@
-package com.controller.cart;
+package com.controller.Cart;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
@@ -7,8 +7,8 @@ import javax.servlet.annotation.*;
 import java.io.IOException;
 import java.util.ArrayList;
 
-import com.dao.cartDao;
-import com.model.cartItem;
+import com.util.CartDaoDatabaseUtil;
+import com.model.CartItem;
 
 @WebServlet(name = "PlusCartItemQuantity", value = "/PlusCartItemQuantity")
 public class PlusCartItemQuantity extends HttpServlet {
@@ -21,7 +21,7 @@ public class PlusCartItemQuantity extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         int id = Integer.parseInt(request.getParameter("id"));
         int cartid = Integer.parseInt(request.getParameter("cartid"));
-        cartDao cart = new cartDao();
+        CartDaoDatabaseUtil cart = new CartDaoDatabaseUtil();
         int res = cart.plusItem(id,cartid);
 
         if(res > 0){
@@ -33,7 +33,7 @@ public class PlusCartItemQuantity extends HttpServlet {
             request.setAttribute("result", "fail");
         }
 
-        ArrayList<cartItem> disCart =  cart.displayCart(cartid);
+        ArrayList<CartItem> disCart =  cart.displayCart(cartid);
         request.setAttribute("cart", disCart);
 
         RequestDispatcher dis = request.getRequestDispatcher("./cart.jsp");
