+++ src/main/java/com/interfaces/Cart.java
@@ -1,0 +1,7 @@
+package com.interfaces;
+import com.model.CartItem;
+import java.util.ArrayList;
+
+public interface Cart {
+    public ArrayList<CartItem> displayCart(int cid);
+}
