--- src/main/java/com/interfaces/cart.java
@@ -1,7 +1,0 @@
-package com.interfaces;
-import com.model.cartItem;
-import java.util.ArrayList;
-
-public interface cart {
-    public ArrayList<cartItem> displayCart(int cid);
-}
