+++ src/main/java/com/model/Cart.java
@@ -1,0 +1,25 @@
+package com.model;
+
+public class Cart {
+
+    private int id;
+    private String cusId;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getCusId() {
+        return cusId;
+    }
+
+    public void setCusId(String cusId) {
+        this.cusId = cusId;
+    }
+
+
+}
