+++ src/main/java/com/model/CartItem.java
@@ -1,0 +1,59 @@
+package com.model;
+
+public class CartItem {
+    private int id;
+    private int cartid;
+    private String iname;
+    private int quantity;
+    private float price;
+
+    public float getPrice() {
+        return price;
+    }
+
+    public void setPrice(float price) {
+        this.price = price;
+    }
+
+    public CartItem(int id, int cartid, String iname, int quantity , float price) {
+        this.id = id;
+        this.cartid = cartid;
+        this.iname = iname;
+        this.quantity = quantity;
+        this.price = price;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getCartid() {
+        return cartid;
+    }
+
+    public void setCartid(int cartid) {
+        this.cartid = cartid;
+    }
+
+    public String getIname() {
+        return iname;
+    }
+
+    public void setIname(String iname) {
+        this.iname = iname;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+
+}
