+++ src/main/java/com/model/Item.java
@@ -1,0 +1,78 @@
+package com.model;
+
+public class Item {
+    private int id;
+    private String name;
+    private String des;
+    private float price;
+    private String category;
+    private String image;
+
+    public String getImage() {
+        return image;
+    }
+
+    public void setImage(String image) {
+        this.image = image;
+    }
+
+    public Item(int id, String name, String des, float price, String category , String image) {
+        this.id = id;
+        this.name = name;
+        this.des = des;
+        this.price = price;
+        this.category = category;
+        this.image = image;
+    }
+
+    private String password;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDes() {
+        return des;
+    }
+
+    public void setDes(String des) {
+        this.des = des;
+    }
+
+    public float getPrice() {
+        return price;
+    }
+
+    public void setPrice(float price) {
+        this.price = price;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+}
