--- src/main/java/com/model/cart.java
@@ -1,25 +1,0 @@
-package com.model;
-
-public class cart {
-
-    private int id;
-    private String cusId;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getCusId() {
-        return cusId;
-    }
-
-    public void setCusId(String cusId) {
-        this.cusId = cusId;
-    }
-
-
-}
